// src/types/requestCriterias.ts
export type DocType = {
  code: string
  label: string
  type: string
}