// src/types/searchCriterias.ts
export enum VitalStatusLabel {
  ALIVE = 'Patients vivants',
  DECEASED = 'Patients décédés',
  ALL = 'Tous les patients'
}

export enum VitalStatusOptions {
  birth = 'birth',
  age = 'age',
  deceasedAge = 'deceasedAge',
  deceasedDate = 'deceasedDate'
}

export enum VitalStatusOptionsLabel {
  birth = 'Date de naissance',
  age = 'Âge actuel',
  deceasedAge = 'Âge au décès',
  deceasedDate = 'Date de décès'
}

export enum VitalStatus {
  ALIVE = 'ALIVE',
  DECEASED = 'DECEASED',
  ALL = 'ALL'
}

export enum Direction {
  ASC = 'asc',
  DESC = 'desc'
}
export enum DirectionLabel {
  ASC = 'Croissant',
  DESC = 'Décroissant'
}
export enum Order {
  ANABIO = 'code-anabio',
  BIRTHDATE = 'birthdate',
  CODE = 'code',
  CREATED = 'created',
  DATE = 'date',
  DESCRIPTION = 'description',
  DIAGNOSIS = 'diagnosis',
  ENCOUNTER = 'encounter',
  FAMILY = 'family',
  FAVORITE = 'favorite',
  FIRSTNAME = 'name',
  GENDER = 'gender',
  ID = 'id',
  IPP = 'identifier',
  LASTNAME = 'lastname',
  LOINC = 'code-loinc',
  MEDICATION_ATC = 'medication_atc',
  MEDICATION_UCD = 'medication_ucd',
  MODALITY = 'modality',
  MODIFIED = 'modified_at',
  NAME = 'name',
  PERIOD_START = 'Period-start',
  PRESCRIPTION_TYPES = 'category-name',
  PROCEDURE = 'procedureCode',
  RECORDED_DATE = 'recorded-date',
  RESULT_SIZE = 'result_size',
  ROUTE = 'route',
  STUDY_DATE = 'started',
  TYPE = 'type-name'
}